+++ chartjs-java-model/src/main/java/software/xdev/chartjs/model/options/IndexAxis.java
@@ -13,28 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package software.xdev.chartjs.model.charts;
+package software.xdev.chartjs.model.options;
 
-import software.xdev.chartjs.model.data.MixedData;
-import software.xdev.chartjs.model.options.Options;
+import java.util.Locale;
+
+import com.fasterxml.jackson.annotation.JsonValue;
 
 
-/**
- * Allows creation of <a href="https://www.chartjs.org/docs/latest/charts/mixed.html">mixed chart types</a>
- */
-public class MixedChart extends AbstractChart<MixedChart, Options<?, ?>, MixedData>
+public enum IndexAxis
 {
-	protected String type;
+	X,
+	Y;
 	
-	public MixedChart setType(final String type)
-	{
-		this.type = type;
-		return this.self();
-	}
-	
+	@JsonValue
 	@Override
-	public String getType()
+	public String toString()
 	{
-		return this.type;
+		return super.toString().toLowerCase(Locale.ENGLISH);
 	}
 }