+++ chartjs-java-model/src/test/java/software/xdev/chartjs/model/BarChartTest.java
@@ -29,6 +29,7 @@ import software.xdev.chartjs.model.data.BarData;
 import software.xdev.chartjs.model.datapoint.XYDataPoint;
 import software.xdev.chartjs.model.dataset.BarDataset;
 import software.xdev.chartjs.model.options.BarOptions;
+import software.xdev.chartjs.model.options.IndexAxis;
 import software.xdev.chartjs.model.options.scale.Scales;
 import software.xdev.chartjs.model.options.scale.cartesian.AbstractCartesianScaleOptions;
 import software.xdev.chartjs.model.options.scale.cartesian.CartesianScaleOptions;
@@ -89,21 +90,6 @@ class BarChartTest extends AbstractChartTest
 	@Test
 	void testStacked()
 	{
-		final BarDataset dataset1 = new BarDataset()
-			.setLabel("First stack")
-			.setData(65, 59, 80)
-			.addBackgroundColors(Color.RED, Color.RED, Color.RED);
-		
-		final BarDataset dataset2 = new BarDataset()
-			.setLabel("Second stack")
-			.setData(28, 45, 10)
-			.addBackgroundColors(Color.GREEN, Color.GREEN, Color.GREEN);
-		
-		final BarData data = new BarData()
-			.addLabels("First", "Second", "Third")
-			.addDataset(dataset1)
-			.addDataset(dataset2);
-		
 		final BarOptions options = new BarOptions()
 			.setAnimation(false);
 		options
@@ -112,7 +98,7 @@ class BarChartTest extends AbstractChartTest
 			.addScale(Scales.ScaleAxis.Y, new CartesianScaleOptions().setStacked(true));
 		
 		this.createScreenshotAndCompare(
-			new BarChart(data, options),
+			new BarChart(createDefaultTwoDatasetBarData(), options),
 			this.getWebContainer(),
 			"Stacked"
 		);
@@ -147,31 +133,16 @@ class BarChartTest extends AbstractChartTest
 	@Test
 	void testStackedHorizontal()
 	{
-		final BarDataset dataset1 = new BarDataset()
-			.setLabel("First stack")
-			.setData(65, 59, 80)
-			.addBackgroundColors(Color.RED, Color.RED, Color.RED);
-		
-		final BarDataset dataset2 = new BarDataset()
-			.setLabel("Second stack")
-			.setData(28, 45, 10)
-			.addBackgroundColors(Color.GREEN, Color.GREEN, Color.GREEN);
-		
-		final BarData data = new BarData()
-			.addLabels("First", "Second", "Third")
-			.addDataset(dataset1)
-			.addDataset(dataset2);
-		
 		final BarOptions options = new BarOptions()
 			.setAnimation(false)
-			.setIndexAxis(BarOptions.IndexAxis.Y);
+			.setIndexAxis(IndexAxis.Y);
 		options
 			.getScales()
 			.addScale(Scales.ScaleAxis.X, new CartesianScaleOptions().setStacked(true))
 			.addScale(Scales.ScaleAxis.Y, new CartesianScaleOptions().setStacked(true));
 		
 		this.createScreenshotAndCompare(
-			new BarChart(data, options),
+			new BarChart(createDefaultTwoDatasetBarData(), options),
 			this.getWebContainer(),
 			"StackedHorizontal"
 		);
@@ -207,7 +178,7 @@ class BarChartTest extends AbstractChartTest
 		
 		final BarOptions options = new BarOptions()
 			.setAnimation(false)
-			.setIndexAxis(BarOptions.IndexAxis.Y);
+			.setIndexAxis(IndexAxis.Y);
 		options
 			.getScales()
 			.addScale(Scales.ScaleAxis.X, new CartesianScaleOptions().setStacked(true))
@@ -218,5 +189,23 @@ class BarChartTest extends AbstractChartTest
 			this.getWebContainer(),
 			"StackedWithGroups"
 		);
+	}
+	
+	static BarData createDefaultTwoDatasetBarData()
+	{
+		final BarDataset dataset1 = new BarDataset()
+			.setLabel("First stack")
+			.setData(65, 59, 80)
+			.addBackgroundColors(Color.RED, Color.RED, Color.RED);
+		
+		final BarDataset dataset2 = new BarDataset()
+			.setLabel("Second stack")
+			.setData(28, 45, 10)
+			.addBackgroundColors(Color.GREEN, Color.GREEN, Color.GREEN);
+		
+		return new BarData()
+			.addLabels("First", "Second", "Third")
+			.addDataset(dataset1)
+			.addDataset(dataset2);
 	}
 }